Add support for adding columns
Also added the FilterFactory class, so that it can be used as a service.
Remove the Filter dependency on Config instead you have the tables(),
fields(), columns() methods to set the filter. The Config can still be
used for encapsulating the data for your filter, but outside the Filter
logic